refactor: move foobar console related code
