+++ src/components/foobar/index.tsx
@@ -4,9 +4,9 @@ import { useEffect } from "react";
 import { FoobarHintWrapper } from "@/components/foobar/styled";
 import { IS_DEV } from "@/config";
 import { FoobarStoreType, useFoobarStore, FOOBAR_PAGES } from "@/domains/Foobar";
+import { logConsoleMessages } from "@/domains/Foobar/console";
 import { Space } from "@/styles/layouts";
 import { LinkTo } from "@/styles/typography";
-import { logConsoleMessages } from "@/utils/console";
 import { useHasMounted } from "@/utils/hooks";
 
 function checkIfAllAchievementsAreDone(completed: FoobarStoreType["foobarData"]["completed"]) {
