+++ src/domains/Foobar/console.ts
